requirements.txt
requests>=2.28.0
netmiko>=4.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
smolagents>=0.0.6
openai>=1.0.0  # Required by OpenRouter

# Dev dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
types-requests>=2.28.0